expand journaling to all changes to metadata dict
